+++ src/screens/SettingsScreen/SettingsScreen.test.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { shallow } from 'enzyme';
 
-import { NavigationData } from '../navigation/AppNavigator/SettingsNavigator';
+import { NavigationData } from '../../navigation/AppNavigator/SettingsNavigator';
 
 import SettingsScreen from './SettingsScreen';
