+++ src/screens/SettingsScreen/index.ts
@@ -1,0 +1,1 @@
+export { default } from './SettingsScreen';
